fix(code-city): fix timeline, add rich metadata, fix int overflow
Timeline fixes:
- Add globalTimeRange to prevent slider jumps when filtering
- Add sliderValue state for stable UI position
- Fix empty states by enforcing 24h minimum window at max slider
- Fix state sync between slider and timestamp calculations

Rich metadata additions:
- Add multiple visualization modes (churn, activity, velocity, recency, size)
- Add height modes (commits, activity, size)
- Add stats panel with hot files, new files, largest files
- Add file detail modal with actions
- Add legend panel with mode selectors
- Enhanced tooltips with net change and churn info

Growth metrics fix:
- Change FileChurn proto fields from int32 to int64 for large repos
- Fix Go reader int64 accumulators
- Replace confusing growth% with net change + velocity metrics
- Clearer velocity mode description

Removes emojis from UI for cleaner look.

+++ internal/repo/reader.go
@@ -431,8 +431,8 @@ func (e *Entry) GetFileChurnMap(ref string, since, until int64) ([]*gitchatv1.Fi
 
 	type churnAcc struct {
 		commits   int
-		additions int
-		deletions int
+		additions int64
+		deletions int64
 		lastMod   int64
 	}
 	m := map[string]*churnAcc{}
@@ -486,7 +486,7 @@ func (e *Entry) GetFileChurnMap(ref string, since, until int64) ([]*gitchatv1.Fi
 					m[f.Name] = acc
 				}
 				acc.commits++
-				acc.additions += int(f.Size) / 40 // rough estimate
+				acc.additions += f.Size / 40 // rough estimate
 				if ts > acc.lastMod {
 					acc.lastMod = ts
 				}
@@ -510,8 +510,8 @@ func (e *Entry) GetFileChurnMap(ref string, since, until int64) ([]*gitchatv1.Fi
 				m[st.Name] = acc
 			}
 			acc.commits++
-			acc.additions += st.Addition
-			acc.deletions += st.Deletion
+			acc.additions += int64(st.Addition)
+			acc.deletions += int64(st.Deletion)
 			if ts > acc.lastMod {
 				acc.lastMod = ts
 			}
@@ -535,8 +535,8 @@ func (e *Entry) GetFileChurnMap(ref string, since, until int64) ([]*gitchatv1.Fi
 		out = append(out, &gitchatv1.FileChurn{
 			Path:           path,
 			CommitCount:    int32(acc.commits),
-			TotalAdditions: int32(acc.additions),
-			TotalDeletions: int32(acc.deletions),
+			TotalAdditions: acc.additions,
+			TotalDeletions: acc.deletions,
 			LastModified:   acc.lastMod,
 			Size:           sizeMap[path],
 		})
